django_experiment/requirements/base.txt
django-debug-toolbar==4.0.0
Django==4.0.3
djangorestframework==3.14.0
psycopg2==2.9.3